zcash_client_memory::proto

Module memwallet

Structs

Account

AccountBirthday

Accounts

Address

ChainState

EphemeralAddress

EphemeralAddressRecord

Memo

MemoryWallet

A serialized zcash wallet state

Note

NoteId

Nullifier

NullifierRecord

records where a nullifier was spent by block height and tx index in that block

OutPoint

RSeed

ReceivedNote

ReceivedNoteSpendRecord

associates a note and a transaction where it was spent

ReceivedTransparentOutput

Recipient

ScanQueueRecord

SentNote

SentNoteRecord

Record storing the sent information for a given note

ShardTree

A shard tree defined by a cap subtree and shard subtrees

TransactionDataRequest

TransactionEntry

TransactionTableRecord

A record storing transaction data in the transaction table

TransparentReceivedOutputRecord

TransparentReceivedOutputSpendRecord

TransparentSpendCacheRecord

TreeCheckpoint

A checkpoint in a shard tree

TreeEndHeightsRecord

Stores the block height corresponding to the last note commitment in a shard as defined by its level and index in the tree

TreeShard

A shard in a shard tree

TxId

Unique identifier for a zcash transaction

TxLocatorRecord

Maps a block height and transaction index to a transaction ID.

TxOut

WalletBlock

Enums

AccountKind

AccountPurpose

PoolType

RSeedType

RecipientType

ScanPriority

Scope

ShieldedProtocol

TransactionDataRequestType

TransactionStatus