zcash_client_memory

Struct MemoryWalletDb

pub struct MemoryWalletDb<P: Parameters> { /* private fields */ }

The main in-memory wallet database. Implements all the traits needed to be used as a backend.

Implementations

impl<P: Parameters> MemoryWalletDb<P>

pub fn encode<B: BufMut>(&self, buf: &mut B) -> Result<(), Error>

Encode a memory wallet db as a protobuf byte buffer Always uses the latest version of the wire protocol

pub fn decode_new<B: Buf>( buf: B, params: P, max_checkpoints: usize, ) -> Result<Self, Error>

Create a mew memory wallet db from a protobuf encoded byte buffer with version awareness

pub fn new_from_proto( proto_wallet: MemoryWallet, params: P, max_checkpoints: usize, ) -> Result<Self, Error>

Build a memory wallet db from protobuf type with version awareness

impl<P: Parameters> MemoryWalletDb<P>

pub fn new(params: P, max_checkpoints: usize) -> Self

pub fn params(&self) -> &P